refactor: Moved evaluateQueries to the {{container-query}} modifier
